Update Bamboo ACT to support execution in Bamboo environment
This commit provides comprehensive enhancements to Bamboo ACT, enabling it to be executed in Bamboo environment. Changes include updating GitHub Authentication to use tokens, disabling unnecessary log hooks, and altering cloneOptions. Additionally, runtime configurations like adding new 'HostExecutorDir' and 'SafeMode' attributes in the Config struct are included. Other specific considerations for Bamboo ecosystem like Bitbucket repositories and placing restrictions on Gitea have also been implemented.

+++ pkg/container/host_environment.go
@@ -32,6 +32,7 @@ type HostEnvironment struct {
 	Workdir   string
 	ActPath   string
 	CleanUp   func()
+	CleanPath bool
 	StdOut    io.Writer
 }
 
@@ -386,7 +387,12 @@ func (e *HostEnvironment) Remove() common.Executor {
 		if e.CleanUp != nil {
 			e.CleanUp()
 		}
-		return os.RemoveAll(e.Path)
+
+		if e.CleanPath {
+			return os.RemoveAll(e.Path)
+		} else {
+			return nil
+		}
 	}
 }
